implemented suggestions for improving par and var tabs
